requirements.txt
crewai
crewai-tools
setuptools
langchain-groq
load-dotenv